+++ pages/SignupPage/SignupPage.html
@@ -13,6 +13,7 @@
         maiúsculas e/ou minúsculas, números e símbolos como traço (-), underscore (_) e ponto (.)">&#63</button>
       </div>
       <sub id="user-error" class="error hide">Usuário inválido</sub>
+      <sub id="user-exists" class="error hide">Usuário já existe, tente novamente!</sub>
       <div class="input-tooltips">
         <input class='input-password' type="password" placeholder="Defina sua senha">
         <button class="tooltip-password"
